wip screensaver

+++ users/jules/shared/services/screensaver.nix
@@ -12,27 +12,51 @@ in
 
   config =
     let
-      lock-with-cbonsai = pkgs.writeShellScriptBin "lock-with-cbonsai" ''
-        ${pkgs.cbonsai}/bin/cbonsai --screensaver --time 0.05 --wait 2 --message "$(fortune)" --multiplier 10 --life 50 &
+      screensaver = pkgs.writeShellScriptBin "screensaver" ''
+        ${pkgs.cbonsai}/bin/cbonsai \
+        --screensaver \
+        --time 0,05 \
+        --wait 2 \
+        --message "$(${pkgs.fortune}/bin/fortune)" \
+        --multiplier 10 \
+        --life 50 \
+      '';
+      pid-file = "/tmp/kitty-screensaver.pid";
+      start-screensaver = pkgs.writeShellScriptBin "start-screensaver" ''
+        ${pkgs.kitty}/bin/kitty \
+          -- ${screensaver}/bin/screensaver > /dev/null &
         pid="$!"
-        ${pkgs.swaylock}/bin/swaylock -f -c 000000
-        kill "$pid"
+        echo "$pid" > ${pid-file}
+        disown
+        exit
+      '';
+      stop-screensaver = pkgs.writeShellScriptBin "stop-screensaver" ''
+        # notify-send Stopping screensaver
+        pid_file="${pid-file}"
+        if [ -f "$pid_file" ]; then
+          # notify-send Welcome back!
+          pid=$(cat "$pid_file")
+          kill "$pid"
+          rm "$pid_file"
+        fi
       '';
     in
     lib.mkIf cfg.enable {
       home.packages = with pkgs; [
-        swaylock
         swayidle
         cbonsai
-        lock-with-cbonsai
+        screensaver
+        start-screensaver
+        stop-screensaver
+        fortune
       ];
-      security.pam.services.swaylock = { };
       services.swayidle = {
         enable = true;
         timeouts = [
           {
-            timeout = 20;
-            command = "${lock-with-cbonsai}/bin/lock-with-cbonsai";
+            timeout = 5;
+            command = "${start-screensaver}/bin/start-screensaver";
+            resumeCommand = "${stop-screensaver}/bin/stop-screensaver";
           }
         ];
       };
